feat: add button hide/show logic

+++ projects/ngx-explorer/src/lib/components/menu-bar/menu-bar.component.ts
@@ -1,4 +1,6 @@
-import { Component, ElementRef, ViewChild, ViewEncapsulation } from '@angular/core';
+import { Component, ElementRef, OnDestroy, ViewChild, ViewEncapsulation } from '@angular/core';
+import { NodeType } from 'ngx-explorer';
+import { Subscription } from 'rxjs';
 import { ExplorerService } from '../../services/explorer.service';
 import { HelperService } from '../../services/helper.service';
 
@@ -8,10 +10,22 @@ import { HelperService } from '../../services/helper.service';
     styleUrls: ['./menu-bar.component.scss'],
     encapsulation: ViewEncapsulation.None
 })
-export class MenuBarComponent {
+export class MenuBarComponent implements OnDestroy {
     @ViewChild('uploader', { static: true }) uploader: ElementRef;
 
-    constructor(private explorerService: ExplorerService, private helperService: HelperService) { }
+    canDownload = false;
+    canDelete = false;
+    canRename = false;
+
+    private sub = new Subscription();
+
+    constructor(private explorerService: ExplorerService, private helperService: HelperService) {
+        this.sub.add(this.explorerService.selectedNodes.subscribe(n => {
+            this.canDownload = n.filter(x => x.type === NodeType.File).length === 1;
+            this.canDelete = n.length > 0;
+            this.canRename = n.length === 1;
+        }));
+    }
 
     createFolder() {
         const name = prompt("Enter new folder name");
@@ -36,7 +50,9 @@ export class MenuBarComponent {
     }
 
     remove() {
-        this.explorerService.remove();
+        if (confirm("Are you sure you want to delete the selected files?")) {
+            this.explorerService.remove();
+        }
     }
 
     openUploader() {
@@ -50,6 +66,10 @@ export class MenuBarComponent {
 
     download() {
         this.explorerService.download();
+    }
+
+    ngOnDestroy() {
+        this.sub.unsubscribe();
     }
 
 }
